giphy: Refactor Giphy integration to centralize common parts.
This is a prep commit meant to extract and refactor
some parts of the GIPHY integration in common modules
and classes to later use along with the Tenor
integration.
