+++ tools/test-js-with-node
@@ -113,8 +113,8 @@ EXEMPT_FILES = make_set(
         "web/src/fetch_status.ts",
         "web/src/flatpickr.ts",
         "web/src/gear_menu.ts",
+        "web/src/gif_state.ts",
         "web/src/giphy.ts",
-        "web/src/giphy_state.ts",
         "web/src/global.ts",
         "web/src/group_permission_settings.ts",
         "web/src/group_setting_pill.ts",
