+++ web/src/click_handlers.ts
@@ -772,7 +772,7 @@ export function initialize(): void {
     function handle_compose_click(e: JQuery.ClickEvent): void {
         const $target = $(e.target);
         // Emoji clicks should be handled by their own click handler in emoji_picker.js
-        if ($target.is(".emoji_map, img.emoji, .drag, .compose_gif_icon")) {
+        if ($target.is(".emoji_map, img.emoji, .drag, .compose-gif-icon-giphy")) {
             return;
         }
