+++ web/src/gif_state.ts
