+++ web/templates/gif_picker_ui.hbs
@@ -1,10 +1,10 @@
-<div id="giphy_grid_in_popover">
+<div class="gif-grid-in-popover">
     <div class="arrow"></div>
     <div class="popover-inner">
         <div class="popover-filter-input-wrapper">
-            <input type="text" id="giphy-search-query" class="popover-filter-input filter_text_input" autocomplete="off" placeholder="{{t 'Filter' }}" autofocus/>
+            <input type="text" id="gif-search-query" class="popover-filter-input filter_text_input" autocomplete="off" placeholder="{{t 'Filter' }}" autofocus/>
         </div>
-        <div class="giphy-scrolling-container" data-simplebar data-simplebar-tab-index="-1">
+        <div class="gif-scrolling-container" data-simplebar data-simplebar-tab-index="-1">
             {{! We need a container we can replace
             without removing the simplebar wrappers.
             We replace the `giphy-content` when
